// src/types/funcionarios.ts
export interface DadosFuncionarios {
    nome: string; 
    cpf: string;
    email: string;
    cargo: string;
    departamento: string; 
    ativo: Boolean;       
}

export interface Funcionario {
    id: number;
    nome: string; 
    cpf: string;
    email: string;
    cargo: string;
    departamento: string;
    ativo: Boolean;
}